[refactor] moving piecewise_approximation from ml_tanh to core.approximation
